+++ src/openassetio-core/include/openassetio/managerApi/ManagerInterface.hpp
@@ -582,18 +582,6 @@ class OPENASSETIO_CORE_EXPORT ManagerInterface {
   using ResolveSuccessCallback = std::function<void(std::size_t, const TraitsDataPtr&)>;
 
   /**
-   * Callback signature used for an unsuccessful entity resolution.
-   *
-   * This must be called for errors that are specific to a particular
-   * reference in a batch. Exceptions can be thrown to indicate a
-   * whole-batch error.
-   *
-   * The appropriate error code should be used for these errors. See
-   * @fqref{BatchElementError.ErrorCode} "ErrorCode".
-   */
-  using ResolveErrorCallback = std::function<void(std::size_t, const BatchElementError&)>;
-
-  /**
    * Provides a @fqref{TraitsData} "TraitsData"
    * populated with the available data for the requested set of
    * traits for each given @ref entity_reference.
@@ -678,7 +666,7 @@ class OPENASSETIO_CORE_EXPORT ManagerInterface {
   virtual void resolve(const EntityReferences& entityReferences, const trait::TraitSet& traitSet,
                        const ContextConstPtr& context, const HostSessionPtr& hostSession,
                        const ResolveSuccessCallback& successCallback,
-                       const ResolveErrorCallback& errorCallback) = 0;
+                       const BatchElementErrorCallback& errorCallback) = 0;
 
   /// @}
  protected: